requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
gunicorn==20.1.0
tensorflow==2.16.1
opencv-python-headless==4.7.0.72
scikit-learn==1.2.2
pandas==2.0.1
numpy==1.24.3
cloudinary==1.32.0
sqlalchemy==2.0.15
python-dotenv==1.0.0
pytest==7.3.1
pydantic==1.10.7
python-multipart==0.0.6
httpx==0.24.1
joblib==1.2.0
pillow==9.5.0
redis==4.5.5
mlflow==2.3.2
prometheus-client==0.16.0